Refactored relVisitor into separate dataManager #3
